suppression des fichiers o et d deposé par erreur. Ajout de l'ecouteI2C. Ajout de wait dans MD25. encore quelques int64_t remplacé par des int32_t.

+++ asservissement/asserv/codeurs/CodeursDirects.cpp
@@ -39,7 +39,7 @@ void CodeursDirects::getCounts(int32_t* countG, int32_t* countD)
 	// On ajuste le sens des codeurs en cas de problème de cablage
 	if (Config::swapCodeurs)
 	{ //inversion des deux codeurs
-		int64_t temp = *countG;
+		int32_t temp = *countG;
 		*countG = *countD;
 		*countD = temp;
 	}
